Change: sendNotifications display via toasts
Rather than displaying sendNotifications in the titlebar, use toasts.
This allows the user to clear the notification once read and frees
up the titlebar space.
