+++ packages/studio-base/src/Workspace.tsx
@@ -30,7 +30,6 @@ import LayoutBrowser from "@foxglove/studio-base/components/LayoutBrowser";
 import messagePathHelp from "@foxglove/studio-base/components/MessagePathSyntax/index.help.md";
 import { useMessagePipeline } from "@foxglove/studio-base/components/MessagePipeline";
 import MultiProvider from "@foxglove/studio-base/components/MultiProvider";
-import NotificationDisplay from "@foxglove/studio-base/components/NotificationDisplay";
 import PanelLayout from "@foxglove/studio-base/components/PanelLayout";
 import PanelList from "@foxglove/studio-base/components/PanelList";
 import PanelSettings from "@foxglove/studio-base/components/PanelSettings";
@@ -56,7 +55,6 @@ import useNativeAppMenuEvent from "@foxglove/studio-base/hooks/useNativeAppMenuE
 import welcomeLayout from "@foxglove/studio-base/layouts/welcomeLayout";
 import { PlayerPresence } from "@foxglove/studio-base/players/types";
 import { isNonEmptyOrUndefined } from "@foxglove/studio-base/util/emptyOrUndefined";
-import inAutomatedRunMode from "@foxglove/studio-base/util/inAutomatedRunMode";
 
 const log = Log.getLogger(__filename);
 
@@ -429,9 +427,6 @@ export default function Workspace(props: WorkspaceProps): JSX.Element {
             <TruncatedText>{currentSourceName ?? "Foxglove Studio"}</TruncatedText>{" "}
           </SToolbarItem>
           <div style={{ flexGrow: 1 }} />
-          <SToolbarItem style={{ marginRight: 5 }}>
-            {!inAutomatedRunMode() && <NotificationDisplay />}
-          </SToolbarItem>
         </Toolbar>
         <Sidebar
           items={SIDEBAR_ITEMS}
